Tidied up type conversion.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@778 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ radtran/spec_data/loadbuffer.f
@@ -218,7 +218,7 @@ C NOTE: SBLIN is the correction to air broadening so that zero is valid
       NXTREC=IREC
 
 101   CONTINUE
-      NBINX = 1+ (VMAX-VBOT)/WING 
+      NBINX = 1+ INT((VMAX-VBOT)/WING) 
       DO I=1,NBINX
 C       PRINT*,I,VBOT+(I-1)*WING,FSTLIN(IB,I),LSTLIN(IB,I)
        IF(FSTLIN(IB,I).GT.0)LASTBIN(IB)=I
